ngl: Turn off depth test
Usually, the GdkGLContext will have the depth buffer disabled, so the
depth test will not work. However, if the GDK backend does not provide a
config without depth buffer, there will be one.

And if we enable it, it'll do the wrong thing.

For example, it'll paint the wrong thing in the cube flip animation of
GtkStack (like on widget-factory page 2).