Port GError's description to Markdown
The documentation in GLib does not follow the "use the type name as the
section name" pattern, as GLib does not have that many types, and
instead it follows a more classical "put a bunch of related functions
into an index" approach, and add a description to it.

The equivalent approach in a gi-docgen world is to write those long
descriptions as extra content, using Markdown files.

The GError documentation is a prime candidate for this approach: it's
related to GError, but it contains much more information that would not
really fit into the description of the GError type.