Redo the conversion setup for memory formats
The two changes have to goal of limiting the
amount of conversion functions we have to write
and to keep the matrix from growing too big.

The first change is to pass the formats to the
conversion functions, which will let them 'chain'
internally (which is not very efficient, but
can be of use).

The second change is to make the matrix sparse,
since we are only interested in fully supporting
a few target formats.