wip: ngl: Try to fix opacity handling
We can't put alpha in place before doing the offscreen
rendering, or it will affect that rendering, in addition
to being applied by the blit.

I wonder if we should stop carrying around alpha as
global state altogether. What is that needed for?

This needs investigation of cases where we can avoid
the offscreen.

Fixes: #4261