gtkwindow: Change present documentation
The way present is used doesn't line up with the ideal
use case, and in most instances a developer won't
actually have a timestamp to pass. `present()` is
what should be used for cases when you don't have a
timestamp, but the current docs have caused
a lot of confusion.

Instead of saying `present()` shouldn't be used,
say when it should be used since there are still
valid uses.